+++ Src/ImplicitNullability.Plugin.Tests/test/data/UnitTests/ImplicitNullabilityConfigurationEvaluatorTest/MultipleAttributes.cs
@@ -1,0 +1,7 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyMetadata("ImplicitNullability.AppliesTo", "InputParameters, RefParameters, OutParametersAndResult")]
+
+// There attributes are ignored:
+[assembly: AssemblyMetadata("ImplicitNullability.AppliesTo", "OutParametersAndResult")]
+[assembly: AssemblyMetadata("ImplicitNullability.AppliesTo", "InputParameters")]