+++ Src/ImplicitNullability.Plugin.Tests/test/data/UnitTests/ImplicitNullabilityConfigurationEvaluatorTest/NonCompilingAttribute1.cs
@@ -1,0 +1,3 @@
+﻿
+
+[assembly: AssemblyMetadata("ImplicitNullability.AppliesTo", "InputParameters, RefParameters, OutParametersAndResult")]