+++ History.md
@@ -1,4 +1,5 @@
 ### vNext ###
+- Implicit Nullability support for compiled assemblies with [AssemblyMetadata("ImplicitNullability.AppliesTo", "...")] configuration attribute
 - New "Implicit NotNull element cannot override CanBeNull in base type, nullability should be explicit" warning for return values and out parameters
 - New "Implicit NotNull result or out parameter overrides unknown nullability of external code" hint
 - Added suppression of "Base declaration has the same annotation" highlighting on implicitly nullable elements